implements streamed response with scala default ws response is now public however the scala counterpart was public aswell everything else is fine, since the java StreamedResponse had a private header so no breaking change
